TON remnants deleted template pages moved to separate folder NavBar added to Root component
