+++ miniapp/package.json
@@ -17,7 +17,6 @@
   "dependencies": {
     "@telegram-apps/sdk-react": "^2.0.20",
     "@telegram-apps/telegram-ui": "^2.1.8",
-    "@tonconnect/ui-react": "^2.0.5",
     "dotenv": "^16.4.7",
     "eruda": "^3.0.1",
     "react": "^18.2.0",
