+++ miniapp/src/components/Root.tsx
@@ -1,8 +1,6 @@
-import { TonConnectUIProvider } from '@tonconnect/ui-react';
-
 import { App } from '@/components/App.tsx';
 import { ErrorBoundary } from '@/components/ErrorBoundary.tsx';
-import { publicUrl } from '@/helpers/publicUrl.ts';
+import {NavBar} from "@/components/NavBar/NavBar.tsx";
 
 function ErrorBoundaryError({ error }: { error: unknown }) {
   return (
@@ -24,11 +22,8 @@ function ErrorBoundaryError({ error }: { error: unknown }) {
 export function Root() {
   return (
     <ErrorBoundary fallback={ErrorBoundaryError}>
-      <TonConnectUIProvider
-        manifestUrl={publicUrl('tonconnect-manifest.json')}
-      >
         <App/>
-      </TonConnectUIProvider>
+        <NavBar/>
     </ErrorBoundary>
   );
 }
