+++ miniapp/src/pages/template_pages/InitDataPage.tsx
